+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/model/interop/CatalogueDispatcherInterface.java
@@ -63,4 +63,6 @@ public interface CatalogueDispatcherInterface<T> {
     String getShortNames();
 
     String getLinkers();
+
+    String getUniqueMap();
 }